Bug: Fixed multiprocessed lemmatization

+++ src/embedding_search.py
@@ -2,9 +2,12 @@ import os
 import json
 import numpy as np
 import pandas as pd 
+import csv
 
 from paths import *
+from time import time
 from tqdm import tqdm
+from datetime import datetime
 from annoy import AnnoyIndex
 from preprocessing import lemmatize
 from multiprocessing import Pool, cpu_count
@@ -50,10 +53,11 @@ def create_index(embeddings, file_name = None, save_path= None):
 
 def call_lemmatize(data_series, dataframe, file_name = None):
     if file_name: 
-        print(f"Lemmatization started for {file_name}.")
+        print(f"{datetime.now()}: Lemmatization started for {file_name}.")
     else: 
-        print("Lemmatized started for unnamed file.")
-    num_cores = 8
+        print(f"{datetime.now()}: Lemmatized started for unnamed file.")
+
+    num_cores = 60
 
     pool = Pool(num_cores)
     results = pool.imap(func = lemmatize, 
@@ -64,11 +68,11 @@ def call_lemmatize(data_series, dataframe, file_name = None):
     pool.join()
 
     if file_name: 
-        print(f"Lemmatization complete for {file_name}.")
+        print(f"{datetime.now()}: Lemmatization complete for {file_name}.")
     else: 
-        print("Lemmatized complete for unnamed file.")
+        print(f"{datetime.now()}: Lemmatized complete for unnamed file.")
 
-    dataframe["lemmatized_selftext"] = np.array(list(results))
+    dataframe["lemmatized_selftext"] = np.array([result for result in results])
 
 
 def main():
@@ -90,25 +94,39 @@ def main():
     #     query_search_index = AnnoyIndex(query_embeddings.shape[1], 'manhattan')
     
 
-    csv_directory = os.listdir(pjoin(DECOMPRESSED_SUMBISSIONS, "csv"))
-
+    csv_directory = os.listdir(pjoin(DECOMPRESSED_SUMBISSIONS))
+    csv_directory.sort()
 
     for file_name in csv_directory:
-        file_path = pjoin(DECOMPRESSED_SUMBISSIONS, "csv", file_name)
+        file_path = pjoin(DECOMPRESSED_SUMBISSIONS, file_name)
         if (not isdir(file_path)):
-            lemmatized_file_path = pjoin(DECOMPRESSED_SUMBISSIONS, "csv", file_name.split(".")[0] + "_lemmatized.csv")
+            lemmatized_file_path = pjoin(DECOMPRESSED_SUMBISSIONS, file_name.split(".")[0] + "_lemmatized.csv")
 
-            if not os.path.isfile(lemmatized_file_path):  
-                print(f"Processing file {file_name}.")
+            if ((not os.path.isfile(lemmatized_file_path)) and  'lemmatized' not in file_name): 
+                st = time() 
+                print(f"{datetime.now()}: Processing file {file_name}.")
 
-                df = pd.read_csv(file_path, usecols=["id", "selftext", "title", "subreddit", "permalink", "url"])
-                call_lemmatize(data_series=df["selftext"], dataframe=df, file_name = file_name)
+                df_chunks= pd.read_csv(file_path, 
+                                       usecols= ['id', 'selftext', 'title', 'subreddit', 'permalink', 'url'], 
+                                       chunksize=100_000)    
+
+                # df = pd.read_csv(file_path, usecols=["id", "selftext", "title", "subreddit", "permalink", "url"])
+                chunk_array = []
+                for index,chunk in enumerate(df_chunks):
+                    call_lemmatize(data_series=chunk["selftext"], dataframe=chunk, file_name = f"{file_name}, chunk_{index}")
+                    chunk_array.append(chunk)
+                    
+                df = pd.DataFrame()
+
+                for chunk in chunk_array: 
+                    df = pd.concat([df, chunk])
+
                 df.to_csv(lemmatized_file_path, index=False)
                 
-                print(f"File {file_name} processed.")
+                print(f"{datetime.now()}: File {file_name} processed. Took {(time() - st)//60} minutes for {df.shape}")
             else: 
-                print(f"File {file_name} already lemmatized.")
+                print(f"{datetime.now()}: File {file_name} already lemmatized.")
 
-        
+    
 if __name__ == "__main__":
     main()
